@@ -5,44 +5,43 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --prod",
+    "build": "ng build --prod --source-map",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e",
-    "build-prod": "node --max_old_space_size=4096 ./node_modules/@angular/cli/bin/ng build --prod --aot"
+    "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.4",
-    "@angular/cdk": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/material": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@covalent/core": "^1.0.0",
-    "@ngrx/effects": "^5.1.0",
-    "@ngrx/entity": "^5.1.0",
-    "@ngrx/router-store": "^5.0.1",
-    "@ngrx/store": "^5.1.0",
-    "@ngrx/store-devtools": "^5.1.0",
+    "@angular/animations": "^6.0.3",
+    "@angular/cdk": "^6.0.0",
+    "@angular/common": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/core": "^6.0.3",
+    "@angular/forms": "^6.0.3",
+    "@angular/http": "^6.0.3",
+    "@angular/material": "^6.0.0",
+    "@angular/platform-browser": "^6.0.3",
+    "@angular/platform-browser-dynamic": "^6.0.3",
+    "@angular/router": "^6.0.3",
+    "@covalent/core": "2.0.0-beta.2",
+    "@ngrx/effects": "^6.0.1",
+    "@ngrx/router-store": "^6.0.1",
+    "@ngrx/store": "^6.0.1",
+    "@ngrx/store-devtools": "^6.0.1",
     "@types/lodash": "^4.14.104",
     "core-js": "^2.4.1",
     "hammerjs": "^2.0.8",
     "lodash": "^4.17.5",
     "moment": "^2.20.1",
     "ngrx-store-freeze": "^0.2.1",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.2.0",
     "zone.js": "^0.8.19"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.8",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.6.6",
+    "@angular/cli": "^6.0.7",
+    "@angular/compiler-cli": "^6.0.3",
+    "@angular/language-service": "^6.0.3",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -58,6 +57,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.6.2"
+    "typescript": "~2.7.2"
   }
 }
