@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 import { Proveedor } from '../../models/proveedor';
 import { ProveedoresService } from '../../services';
