@@ -1,7 +1,7 @@
+
+import {throwError as observableThrowError,  Observable } from 'rxjs';
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams, HttpHeaders } from '@angular/common/http';
-
-import { Observable } from 'rxjs/Observable';
 import { map, switchMap, catchError } from 'rxjs/operators';
 
 import * as _ from 'lodash';
@@ -20,26 +20,26 @@ export class ClasesService {
   list(): Observable<Clase[]> {
     return this.http
       .get<Clase[]>(this.apiUrl)
-      .pipe(catchError((error: any) => Observable.throw(error)));
+      .pipe(catchError((error: any) => observableThrowError(error)));
   }
 
   save(clase: Clase): Observable<Clase> {
     return this.http
       .post<Clase>(this.apiUrl, clase)
-      .pipe(catchError(error => Observable.throw(error)));
+      .pipe(catchError(error => observableThrowError(error)));
   }
 
   update(clase: Clase): Observable<Clase> {
     const url = `${this.apiUrl}/${clase.id}`;
     return this.http
       .put<Clase>(url, clase)
-      .pipe(catchError(error => Observable.throw(error)));
+      .pipe(catchError(error => observableThrowError(error)));
   }
 
   delete(clase: Clase) {
     const url = `${this.apiUrl}/${clase.id}`;
     return this.http
       .delete(url)
-      .pipe(catchError(error => Observable.throw(error)));
+      .pipe(catchError(error => observableThrowError(error)));
   }
 }
