@@ -1,9 +1,7 @@
+import { throwError as observableThrowError, Observable, of } from 'rxjs';
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams, HttpHeaders } from '@angular/common/http';
-
-import { Observable } from 'rxjs/Observable';
 import { catchError } from 'rxjs/operators';
-import { of } from 'rxjs/observable/of';
 
 import * as _ from 'lodash';
 
@@ -25,7 +23,7 @@ export class LineasService {
     });
     return this.http
       .get<Linea[]>(this.apiUrl, { params: params })
-      .pipe(catchError((error: any) => Observable.throw(error)));
+      .pipe(catchError((error: any) => observableThrowError(error)));
   }
 
   get(id: string): Observable<Linea> {
@@ -36,7 +34,7 @@ export class LineasService {
   save(linea: Linea): Observable<Linea> {
     return this.http
       .post<Linea>(this.apiUrl, linea)
-      .pipe(catchError(error => Observable.throw(error)));
+      .pipe(catchError(error => observableThrowError(error)));
   }
 
   update(linea: Linea): Observable<Linea> {
@@ -48,6 +46,6 @@ export class LineasService {
     const url = `${this.apiUrl}/${id}`;
     return this.http
       .delete(url)
-      .pipe(catchError(error => Observable.throw(error)));
+      .pipe(catchError(error => observableThrowError(error)));
   }
 }
